Filters signaled events out of wait list.
This change filters already signaled events out of the wait lists
used when enqueueing level zero commands into command lists. This
fixes an issue where an already waited/signaled event can appear
in a wait list in a command batch, and before the batch is
submitted a piEventRelease call can remove the ZeEvent that is
being referred to in the bacth's command list.