[Driver] Use clang-offload-deps tool for generating device dependencies
This patch changes driver to use clang-offload-deps tool for generating
dependence objects for SYCL and OpenMP programs when static fat libraries
are used. Driver adds an additional link step on the host side to create
host image which serves as an input for the clang-offload-deps tool.
Dependence bitcode files produced by the tool are then compiled to
objects where necessary and added to the device linker inputs.

Signed-off-by: Sergey Dmitriev <serguei.n.dmitriev@intel.com>