Add clang-offload-deps tool for generating dependence files for offload targets (#2872)
This tool is intended to be used by the clang driver for offload linking with
static offload libraries. It takes linked host image as input and produces
bitcode files (one per offload target) containing references to symbols that
must be defined in the target images. Each dependence bitcode file then is
expected to be compiled to an object by the driver using appropriate target
toolchain, and dependence object added to the target linker as input together
with the other inputs. References to the symbols in dependence object should
ensure that target linker pulls in necessary symbol definitions from the input
static libraries.

Signed-off-by: Sergey Dmitriev <serguei.n.dmitriev@intel.com>