[llvm-link] fix linker behavior when linking archives with --only-needed option
This patch fixes linker behavior when archive is linked with other inputs as
a library (i.e. when --only-needed option is specified). In this library is
expected to be normally linked first into a separate module and only after
that linker should import required symbols from the linked library module.

This is the same change as https://reviews.llvm.org/D92535

Signed-off-by: Sergey Dmitriev <serguei.n.dmitriev@intel.com>