[SYCL][L0] Add support for pinned host memory. (#2633)
This change implements the support for pinned host memory in level_zero plugin. It allocates pinned host memory when PI_MEM_FLAGS_HOST_PTR_ALLOC is set. Pinned host memories are automatically accessible from the device through PCI. This property also ensures that the memory map/unmap operations are free of cost and the buffer is optimized for frequent accesses from the host.

A test is added for this change in a separate PR at https://github.com/intel/llvm-test-suite/pull/83

Signed-off-by: rbegam <rehana.begam@intel.com>