Adds an environment variable to control event wait list filtering.
This edit adds an environment variable, currently intended for
internal use only (and therfore not documented), that can be used
to control whether already signalled events are filtered out of
the command event wait lists. The environment variable is
SYCL_PI_LEVEL_ZERO_FILTER_EVENT_WAIT_LIST, and if set causes the
filtering to occur.