[SYCL] Enable host optimization of work-item free functions (#2967)
The SYCL free functions ( this_item, this_id, etc) are expensive to support on host devices. They cause performance delays because every iteration through one of the parallel_for routines the various indexing values have to be updated in case the users code might call this_item or this_id (or the others). But with the new callsThisItem method added to the Kernel Information, the host device can avoid paying the performance penalty if the users code doesn't actually call this_item. We can detect at compile time whether or not any of the this_xxx free functions are used by the users code, and if not, don't bother storing the indexing data in each loop iteration.

In this PR we add further expand the Kernel Information to support a callsAnyThisFreeFunction method, and we use it to avoid the sundry store_item etc. calls on the host.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>