Make Intel attributes consistent with clang attributes.
Clang allows attributes to be specified via a special pragma named
`#pragma clang attribute` and many of the Intel-specified attributes
were opting out of this support for no real reason.

This changes allows Intel attributes to be specified via this pragma,
similar to how almost all other Clang attributes are.