[SYCL] Put constant initializer list data in non-generic addr space. (#3005)
* [SYCL] Put constant initializer list data in non-generic addr space.

Big constant initializer lists (>= 16 elements) were represented as a
global in the generic address space. This is incorrect, as this is an
alias AS, and allocation is not possible in it - compiler will have no
clue where to really put data. Gen vector BE crashed on this pattern.
The fix is to put this data into the same address space as a string
literal.

Signed-off-by: Konstantin S Bobrovsky <konstantin.s.bobrovsky@intel.com>