Fixes a deadlock that can occur with batching enabled.
This change causes the event wait list for any command being
enqueued to be checked to see if it was issued into a different
queue than the command that will be waiting on it.  If it was,
then for batching to work properly, the currently open batch for
that queue is closed and submitted for execution.