[SYCL] Change the device index number to start at one.
In recent discussion, we decided to change the definition of device number in SYCL_DEVICE_FILTER.
Originally, it was the enumeration order from get_devices() API.
But, it is better to use the sycl-ls enumeration order, which actually uses the same device enumeration order.
Further, we also decided to start the device index at one instead of zero for user-friendliness.

Signed-off-by: Byoungro So <byoungro.so@intel.com>