[SYCL] Refactor the way we diagnose some SYCL attributes; mostly NFC (#3147)
This removes a helper method from some SYCL attributes which would
provide the name of the attribute. This code wasn't needed because the
diagnostics engine already knows how to print attributes by name.

While making this change, some diagnostics are now properly quoted when
they weren't previously. As a drive-by, I also noticed that the location
of the diagnostics was wrong (it was diagnosing the start of the
attribute range rather than the problematic attribute). This means that
duplicate or conflicting attributes should be warning about the second
attribute (the one causing the conflict) rather than the first
attribute.