[SYCL] Switch from COPY_HOST_PTR to write for devices w/o HUM (#3105)
Switch from initializing native memory objects with COPY_HOST_PTR to
creating them without a host ptr and following up with a write
operation.

The reasoning behind this change is that some OpenCL runtimes (e. g.
FPGA) use the queue the memory object write is enqueued to as a hint to
which device from the context this memory should be copied to, as
opposed to COPY_HOST_PTR where only the context itself is known. This
led to unnecessary temporary host copy made by the FPGA runtime during
buffer creation.