Remove a parsing error and make it a semantic error
The SYCL loop attributes can only be applied to a for, do, or while
statement. However, we would diagnose this as a parsing error rather
than a semantic error. Attribute appertainment of [[]] attributes is
handled as a semantic error to avoid potentially changing the parsing
paths in the presence of an [[]] attribute.

As part of this change, I also corrected the attribute definitions to
properly reflect that these are statement attributes and have a
specific subject list. However, Clang's automated diagnostics focus
primarily on declaration attributes, so this change is mostly for
code readability rather than correctness (for the moment).