[SYCL] Store expression as ConstantExpr in Semantic Attribute
Currently , the behavior is inconsistent for FPGA attributes. We store
the ConstantExpr for some SYCL/FPGA attributes (not all).  This patch is
part 1 of making it consistent. It modfies
addIntelSYCLSingleArgFunctionAttr() to save ConstantExpr. This saves
ConstantExpr for all attributes that calls this function. A follow-up
patch will address the remaining attributes.

Signed-off-by: Elizabeth Andrews <elizabeth.andrews@intel.com>