[SYCL] Temporarily restore event deps in in-order queues for Level Zero
Some of the enqueued operations are implemented directly in the Level
Zero plugin (e.g. map/unmap) and their implementation does not satisfy
in-order queue requirements, leading to incorrect behaviour if the event
dependencies that should be redundant are dropped.

Restore these event dependencies until this is fixed in the Level Zero
plugin.