[SYCL] Store expression as ConstantExpr in Semantic Attribute (#3169)
Currently , the behavior is inconsistent for FPGA attributes. We evaluate and store the ConstantExpr for some SYCL/FPGA attributes , and discard it for others. This patch is part 1 of making the behavior consistent. This patch saves the ConstantExpr for the following attributes. The ConstantExpr is then extracted from the Sematic attributes and used in CodeGen instead of reevaluating it.

SYCLIntelNoGlobalWorkOffsetAttr,
SYCLIntelMaxGlobalWorkDimAttr,
SYCLIntelNumSimdWorkItemsAttr ,
IntelReqdSubGroupSizeAttr ,
SYCLIntelSchedulerTargetFmaxMhzAttr ,
ReqdWorkGroupSizeAttr,
IntelReqdSubGroupSizeAttr,
SYCLIntelMaxWorkGroupSizeAttr

A consequence of this change is a change in diagnostic when the attribute argument is not an integer constant expression. The new diagnostic is an existing llvm-project diagnostic.

A follow-up patch(s) will address the remaining attributes.

Signed-off-by: Elizabeth Andrews elizabeth.andrews@intel.com