Disable Travis CI on master branch
We have GitHub Actions configured and there is no need to use two CI
systems at the same time.

The .travis.yml config file is not removed completed yet as we haven't
swithed to use GitHub Actions on release branches - this file will be
kept as some kind of reference for now.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/06bddc1