Add an attr tablegen feature for checking langopts
For SYCL, we want some attributes to be ignored in host mode, but
without diagnosing the attribute as being ignored. This adds a bit to
the tablegen emitter to specify that a given language option opts into
this behavior. This allows us to more naturally specify that an
attribute is a device-only attribute from Attr.td without adding logic
in SemaDeclAttr.cpp to bail out in host mode (which looks like a bug
due to the lack of an ignored attribute warning).

As a drive-by, this also corrects a think-o with the
[[intel::reqd_sub_group_size]] attribute in OpenCL mode, which was
always being silently dropped.