[NFC][ESIMD] Remove one of the uses on __SYCL_EXPLICIT_SIMD__
This patch is a part of the efforts for allowing ESIMD and
regular SYCL kernels to coexist in the same translation unit and
in the same program.

In ESIMD kernels, calls to __spirv_ functions are lowered into GenX intrinsic.
Because of that, it's OK to have definitions instead of declarations for those functions.
Unused definitions of __spirv_ functions will be DCE'ed anyway, so no impact on functionality.