[ESIMD] Move ESIMD-specific passes to sycl-post-link (#3222)
This patch is a continuation of prior series of patches to allow
having ESIMD and regular SYCL kernels in the same source code
and in the same program.

To mix two types of kernels we need to postpone lowering of
ESIMD specific constructs and do that after we split ESIMD from
regular SYCL code, which is done in sycl-post-link.

This is a finalizing patch that only flips the switch for
doing ESIMD-specific transformations in sycl-post-link
and removes those passes from the common optimization
pipeline in FE (BackendUtil.cpp). All the preliminary patches
have already been committed.