[NFC] Update GlobalOffset pass for new pass manager
Refactored GlobalOffset pass to use new pass manager interface and added
a wrapper to use it with the legacy pass manager.

At the moment this is a non-functional change, because CodeGen passes
are still scheduled through the legacy pass manager and some additional
changes will be needed to include the pass into a pipeline when it will
be composed using the new pass manager.