Rework the way we handle SYCL -cc1 arguments.
-fsycl-is-device and -fsycl-is-host are mutually exclusive options that
now imply -fsycl mode. So passing -fsycl-is-device will now set both
the SYCL and SYCLIsDevice language options.

This introduces driver errors if the user attempts to pass both host
and device mode on the command line. Further, it disallows you from
specifying the -sycl-std= option unless in device or host mode.

There is some testing fallout from these changes because some tests
were passing -fsycl-is-device and not passing -fsycl which meant that
SYCL mode was never enabled for those tests.