[SYCL] Fix  Host Device sampler coordinate conversion (#3226)
The None and Clamp sampler are incorrectly calculating isOutOfRange predicate when the image dimensions are less than 3. This is because the shorthand float conversion used for the samplers uses 0 for the missing dimensions. But that isn't correct, when linearly sampling, the pixel center is at 0.5. Setting it to 0 then leads to an 'out of bounds' for those undefined dimensions. This fix is to use 0.5 for missing dimension.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>