[SYCL] Add ITT annotation instructions
This patch introduces InstrumentalAnnotationsPass pass to sycl-post-link
tool. The pass is being included to sycl-post-link chain only when
'-fsycl_device_code_add_instrumentation_calls' option is passed to
clang driver.

Current version of the pass create instruction to notify profiling tools
that: work item has started/finised or resumed (in case if barrier was
called) and it annotates the barrier call itself and for annotation
of atomic instructions.

Signed-off-by: Dmitry Sidorov <dmitry.sidorov@intel.com>