[SYCL] Create an empty integration header file with -fsycl-int-header (#3213)
When -save-temps is specified on the command-line, the Driver assumes
that an integration header file gets generated and includes the file
for consumption by subsequent phases.  However, currently the file
only gets created when there is something to emit inside it.  This
causes missing include file failures when a later compilation phase
expects this integration header file.

Signed-off-by: Premanand M Rao <premanand.m.rao@intel.com>