[SYCL] Add Level-Zero interop with specification of ownership (#3231)
The Level-Zero backend specified assuming ownership of the native
handles used in the interoperability. There was however request made
that this is not sufficient in all cases, e.g. when some memory was created
outside SYCL RT in some Level-Zero Context, then working with that
memory should always be within the same context. Thus if anyone wanted
to "return" something back from SYCL, then the context should still exist
and be the same.

This patch only does for context, but probably we'd need to deal with other
objects too at some point.

Signed-off-by: Sergey V Maslov sergey.v.maslov@intel.com