[SYCL] Fix a typo in test regression/device_num.cpp
The change is simple.
">=" ==> ">"
The original code was written assuming the device_num starts at 1.
As I changed the code to start device_num at 0, the condition of test
should have been changed to NOT run when the test is going over the total
number of devices found in the system.

Signed-off-by: Byoungro So <byoungro.so@intel.com>