[SYCL] Add SYCL_HOST_UNIFIED_MEMORY environment variable
Add the SYCL_HOST_UNIFIED_MEMORY environment variable that forces the
execution graph builder to behave as if either no non-host devices support
host unified memory or all of them do. The purpose of this variable is to
facilitate testing and debugging of host unified memory specific
behaviour in the runtime library.