[SYCL] Move devicelib req mask collection into a LLVM pass (#3291)
Fallback device library 'on-demand' loading mechanism requires offline
analysis of user's input device image to find out all fallback device
libraries (spv) needed in runtime and the offline analysis is done in
sycl-post-link. Currently, the analysis implementation is in
sycl-post-link.cpp which is not the best practice since sycl-post-link.cpp
should only include the implementation itself. This PR aims to move the
analysis into a LLVM module pass, the sycl-post-link.cpp will on be
responsible for running this pass but includes no actual implementation.

Signed-off-by: gejin <ge.jin@intel.com>