[Driver][SYCL][FPGA] Improve FPGA AOT when using Triple
Behaviors for enabling AOT for FPGA is allowed when using both
-fintelfpga and -fsycl-targets=spir64_fpga-unknown-unknown-sycldevice.
Improve the usage when using the -fsycl-targets method as we were
a bit too reliant on -fintelfpga for some checks, causing some
of the tool chain construction to fail.

Use of -fintelfpga should effectively only enable specific settings
such as debug and dependency generation.