[SYCL] Remove redundant build options processing (#3342)
After 86b0e8d5 patch extra operations with device images happen before
check if it is present in in-memory cache. For applications with small
kernels which are executed multiple times noticeable performance
degradation is seen for host code. That was done to get build options
stored in the kernel image and use them as in-memory cache key. At the
same time kernel image (where these options are taken from) is used in
cache key so it is reasonable to use only build options which are
specified in SYCL API and/or environment variables as separate cache
key.

Getting build options from kernel image is moved back to build
operation which happens only if built program is missed in in-memory
cache.