[ESIMD] Mark ESIMD globals with sycl_explicit_simd attribute (#3348)
[ESIMD] Mark ESIMD globals with sycl_explicit_simd attribute

This patch is a part of the efforts for allowing ESIMD and
regular SYCL kernels to coexist in the same translation unit and
in the same program.

There are differences in semantic analysis of global variables
used in ESIMD and regular SYCL kernels. One such difference is that
ESIMD does not permit private global variable to have an initializer.

Before this patch we used -fsycl-explicit-simd compiler flag to
differentiate globals used in ESIMD and regular kernels. With this
change we mark ESIMD globals with sycl_explicit_simd attribute. So,
the goal of this change is to remove dependency on compiler option,
while preserving the same behavior in semantic analysis.

Co-authored-by: Aaron Ballman <aaron@aaronballman.com>