[Driver][SYCL][FPGA] Support for FPGA device archives created for emulation
The expected input type for AOCX/AOCR were considered for hardware
usage with FPGA.  In this part, we allow for the differentiation of
the type of archive created, whether it be emulation or hardware.
Depending on how these archives are created, we make sure we take
the proper flow to process the contents.

Also perform some cleanup in the needed triple arch representation
for fpga, where we didn't need to define all of the different types
we were wrapping for FPGA specific binaries.