[SYCL] Replace spec constants 2020 with RT-provided values in sycl-post-link
By doing this change, SpecConstantsPass starts to replace scalar specialization
constants with a user-provided value from RT (instead of using C++ default
initialization value) for AOT mode.

This is done with the new intrinsic __sycl_getScalar2020SpecConstantValue:
the 3rd argument `RTBuffer` is a pointer to a runtime buffer, which holds
values of all specialization constants.