[Sema] Introduce deferred diagnostics for coexisting languages (#3395)
Multiple offloading languages can coexist in the same translation
unit, while having different semantics and thus require different
diagnostics. The same construct in the device code may be valid
for one language but invalid for another, while coexisting in a
single translation unit. One such example is SYCL/ESIMD, which
have difference in semantic analysis of globals (will be uploaded
in consequent patches).

This patch introduces a generic mechanism for distinguishing
deferred diagnostics for coexisting languages. The idea of the
mechanism is to 1) extend a deferred diagnostic to store a 'reason'
for creating it and 2) keep track of the funciton context while
emitting a deferred diagnostic.

Example: when a certain expression is allowed only for ESIMD and
is forbidden for SYCL, we store in the diagnostic itself that we
will emit it only for SYCL code. Later when we traverse a call
graph of to-codegen functions we keep track of the function context,
which is determined by the root function. If the context matches
the reason of saving the diagnostic, we emit it, otherwise not.

Co-authored-by: Erich Keane <erich.keane@intel.com>