Get to a reasonable position so that we can pick this up later.
Currently the discussion on the design has lead me to believe that
writing this patch is premature. Some of the in-flight changes to the
spec will severely change what this implementation looks like, so we're
going to put this patch on hold until the spec settles.

This is hopefully in a reasonable enough position where we can pick it
up later without trouble.