[SYCL] Add infrastructure for post-integration header
We need a separate header that the driver will insert at the end of host
translation unit to support SYCL 2020 spec-constants and
is_device_copyable (plus potentially more in the future).  This patch
puts the infrastructure together so that these things can be added when
final specifications are available, and so that the driver can implement
their component.