[SYCL][FPGA] Allow max-concurrency attribute on functions. (#3388)
This patch implements the support of a new FPGA function attribute ‘max_concurrency”. The attribute exists already for loops. It takes a single unsigned integer argument and has the following syntax:
[[intel::max_concurrency(n)]]

An example of it use:
[[intel::max_concurrency(n)]]
void foo() {
}
The LLVM IR representation will be function metadata:
!max_concurrency !0
!0 = !{!i32 n}

Max_concurrency applies to functions in device code. It is not be propagated to the caller.