[Driver][SYCL] Improve integration header usage when performing offload
When generating the integration header, we were performing an additional
compilation step which is not needed, as we can create the output file
and the integration header at the same time.

Improve this behavior by tracking the integration files based on given
input files, creating and using the files accordingly when needed during
the compilation flow.