SYCL: Implement some dependencies for SYCL 2020 Spec constant design
The Spec-constant design (https://github.com/intel/llvm/pull/3331) has a
few dependencies on the CFE, including generating unique names for
reachable specialization_id variables and putting the results in the
integration footer.

This patch helps with this effort in 2 ways:
First, it creates a command line option that will be used eventually to
ensure that types with internal linkage have names unique to this
translation unit.  This will eventually be used by the
unique-id/unique-stable-name implementation, but is required as a CC1
option to unblock the Driver implementation.

Second, it generates the integration footer sans the generated names
(which will be added when we have the unique-id/unique-stable-name
implementation in place).  This is necessary to unblock library
implementation of this feature.