[SYCL] Fix DeclareSYCLBuiltins for round-trip setting. (#3491)
Upstream LLVMadded a round-trip functionality to the command line args so
that we can generate a full set of command line arguments from our
existing state.  At the moment, this is 'assert' build only, but it
seems the intent is to make this perminent for future functionality.

We will have to analyze the full set of our command line args to make
sure we do this right (that is, that it gets set right the 2nd time),
but this one fixes the DeclareSYCLBuiltins to unblock the pulldown's
build.