[SYCL] Correct the tablegen for checking mutually exclusive stmt attrs (#3519)
The work that was done upstream turned out to be insufficient for
checking statement attribute mutual exclusion because attributed
statements do not collect their attributes one-at-a-time in the same
way that declarations do. So the design that was attempting to check
for mutual exclusion as each attribute was processed would not ever
catch a mutual exclusion.

The new approach is to check all of attributes that are to be applied
to the attributed statement in a group. This required generating
another DiagnoseMutualExclusions() function into AttrParsedAttrImpl.inc.