[ESIMD] Skip rewriting functions used through function pointers (#3527)
This fixes a crash when running one of the ESIMD tests that use function pointers with `-O0`. The problem here is that the `ESIMDLowerVecArg` pass expects users of the functions to be only call instructions, but it's not the case with function pointers since the address of a function can be stored in some variable.