[SYCL] Implement anonymous namespace spec-id functionality.
In order to get the specialization_id metadata to be usable in the
integration footer, we need to insert a 'shim' function at each
anonymous namespace so that we can do a lookup at the 'right' location,
then look it up again later by the type.

This implements all that functionality based on my latest understanding
of the design.