[Driver][NFC] Cleanup some option setting for SYCL offload (#3542)
The area in which we setup options for the device and host compilations
is a bit disorganized.  Clean up some of this to be a little easier to
find where additional options should go.  Also clean up the convoluted
usage of UseSYCLTriple which was used largely in part when doing the old
method of static device lib compilations.