[SYCL][PI][L0] Close and submit batch immediately when Queue is empty. (#3552)
This change causes executeCommandList to close a command list and
immediately execute it when it has determined that the last command
submitted to the queue has already been completed.

Updates code to only consider whether the Queue is empty when
doing dynamic batching.  Fixed size batching will not consider this
and give user the exact batching specified.  This change-set also
updates the dynamic batching unit test.