[SYCL] Adds support for device UUID as a SYCL extension.
This includes support for only level_zero. The
llvm/sycl/docs/extensions/IntelGPU/IntelGPUDeviceInfo.md doc has
been modified with a brief description of the extension. A new aspect
is added to indicate if the support is available.

This also fixes the mismatched return types of few other intel specific
ext.

Signed-off-by: rbegam <rehana.begam@intel.com>