[ESIMD] Add support for widening SIMT functions - FE part
This patch is one of the few patches to provide a functionality to reuse
existing scalar function implementations in a vector context.

For example, based on scalar version of `float cos(float)`, implement
simd<float,N> cos(simd<float,N>) - "widened" version of the original
scalar function.

To achieve that, we mark such functions with `sycl_esimd_widen` attribute
which will be later used by vector compiler backend to widen it.

For now, this functionality will be used only in internal implementation of
specific public ESIMD APIs. It is not supposed to be used directly in the
user code. It is a temporary solution to enable ESIMD math functions by
reusing existing scalar implementations. Later it is supposed to be
replaced by a more generic mechanism.