[SYL][PI]{L0]Submit open command batch on event status query
This closes and executes a queue's open command batch list when an
event that was created in the queue has its execution status
queried. This isn't strictly required, but if the execution status
is being queried, it seems like a good idea to make sure any existing
commands have been submitted. This fixes a situation where if you query
over and over on the event's status without ever enqueueing more
commands, or doing a wait, the command may remain unsubmitted.