[SYCL][CUDA] Consider offsets when inferring memcpy pitches
In some situations, a zero row/slice pitch is passed for source and/or
destination to `commonEnqueueMemBufferCopyRect`. In this case, the
pitches are inferred from the copied region. However, if source and/or
destination have offsets in the x or y dimension, those must be taken
into account as well (CUDA checks this, and will otherwise return
`CUDA_ERROR_INVALID_VALUE`).