Add __builtin_num_fields and __builtin_field_type.
This is an experimental branch to see whether we can implement the
device copyability checking functionality within the SYCL library by
using some builtins to provide super rudimentary introspection of
structures and lambdas. These builtins will tell you how many fields
are in the structure or how many captures are in the lambda, and tell
you what types those fields are.

__builtin_num_fields() accepts a type id and returns the number of
fields in the given type as an integer constant expression.
__builtin_field_type() accepts a type id and an integer constant
expression and returns a "value" of the field at the given offset from
the given type. The call should always be wrapped in a decltype (or
similar) expression to get the actual type of the field. The value
returned is meaningless and it is UB to inspect it. Note: if we wish to
upstream any of this functionality, this builtin should be replaced by
a new type in the type system, akin to how decltype and typeof work.

More work is still needed for the SYCL library to fully support the
checking in that there is a need to get the number of base classes and
the base class type information.