Add __builtin_num_bases and __builtin_base_type.
NOTE: this has a fair amount of code duplication that should be cleaned
up if we're happy with the experiment. The num_bases and num_fields AST
nodes are basically the same thing and so those could be differentiated
with a single bit in the AST node. A common base class might also be a
reasonable approach.