[SYCL] Rename helper function to avoid collision
Helper functions have the same names but different signatures to the
ones to be mapped to SPIR-V operations. That makes some backends
generate external calls for them which causes undefined reference
on JIT compilation. The problem is seen only when compiling with
-O0 or -O1 optimization options because when -O3 or -O2 are used
helper functions are inlined and definitions are removed from final
code eliminating the issue.