Fix memory leak after getAsInstruction
Adding a destructor and a special class field to account for getAsInstruction instruction.
Per LLVM documentation, getAsInstruction doesn't attach the result to any Module/Function/Basic Block and therefore it is not deleted as part of their destructors; basically, any pointers to this memory are lost at the end of corresponding if scopes.

Signed-off-by: amochalo <anastasiya.mochalova@intel.com>

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/a05fc6c