Relax handling of boolean specialization constants
When SPIR-V friendly IR is used, it might be the case that it was
generated by some other LLVM tool and/or pass and for boolean types it
is possible that they are represented as i8 instead of i1 in LLVM IR
(depending on how they were used in the original source file).

This patch adjusts and relaxes handling of boolean specialization
constants from SPIR-V friendly IR to get the following signatures
acceptable: `i1 _Z20__spirv_SpecConstantia(i32, i8)`.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/38db037