[SYCL][Doc] Design doc: optional kernel features
First public review of the design document to conform to the SYCL 2020
"optional kernel features" behavior.  This mostly covers SYCL 2020
section 5.7 "Optional kernel features", but it also covers the design
for the `[[sycl::requires()]]` attribute.

Since a SPIR-V extension is proposed as part of this design, this
commit also includes a proposed extension to SPIR-V for enabling
code conditionally based on specialization constants.  The SPIR-V
extension is more general than required for the "optional kernel
features" design because it also includes functionality that will be
required for the SYCL_EXT_ONEAPI_DEVICE_IF extension to DPC++.