[SYCL] Use SYCL-2020 backend_return_t<> in interop_handle instead of interop<>
The corresponding changes in LIT tests:
https://github.com/intel/llvm-test-suite/pull/263

This patch also removes the implementation of the method get_native_mem
accepting accessor with target not equal to global_buffer/constant_buffer.
That method always threw exception, which is wrong as that is the runtime
handling of an obvious error, which must be detected earlier - at compile time.

The template parameter 'backend Backend = backend::opencl' remains as one
having the default value to not break compatibility with older behavior
of interop_handle::get_native_mem(), which could be called without
explicit specification of the 'backend' template parameter.

Signed-off-by: Vyacheslav N Klochkov <vyacheslav.n.klochkov@intel.com>