[SYCL] Implement SYCL2020 functionality: no propagation to the caller
In SYCL 1.2.1 spec, the attributes get propagated from device functions to a kernel.
SYCL 2020 spec does not allow propagating the attributes to the caller when used on a function.

Attributes that should not be propagated from device functions to a kernel to match with new SYCL 2020 spec.

1. scheduler_target_fmax_mhz
2. kernel_args_restrict
3. no_global_work_offset
4. max-work-group-size
5. max-global-work-dim
6. num-simd-work-items
7. reqd-sub-group-size
8. reqd-work-group-size
9. named_sub_group_size

This patch keeps the SYCL 1.2.1 spec functionality and propagates the attributes with the older SYCL mode
(-sycl-std=2017).

Signed-off-by: Soumi Manna <soumi.manna@intel.com>