[SYCL] Fix to alignment assertion (#3671)
There are several reasons sycl::aligned_alloc_host might fail,
for example if there is not enough memory. The present
assertion seems to want to ensure that if the alignment is
used that the memory is, ultimately, properly aligned. But it
is too open and could instead assert if an alignment was
specified and any failure occurred (even one not associated
with alignment). This fix tightens that.

Signed-off-by: Chris Perkins chris.perkins@intel.com